major adjustments to code execution/logic.    (1) subs/occupation_prob.f  — Occupation prob set to 1 if >  0.99999999D0    (2) newsubs/quad_multi_v10.f  — Bug fix. T2 instead of T3 was being compared to PHOT_DIS_PARAMETER.
Changed REAL*8   to REAL(10) and DOUBLE PRECISION to REAL(10).
In some routines it was necessary to change DSQRT to SQRT, DABS to ABS etc.
Primary reason for these changes were to handle the very low temperatures  that were occurring
   in Gamma Velorum models.

In some routines, explicit number calls were replaced. That is 1.0D0 in call
was replaced by RONE etc. Routines changed in this way include:
      blas/dasum.f
      blas/dcabs1.f
      blas/drotg.f
      blas/idamax.f

      disp/maingen.f
      disp/subs  — dlogvec.f, fg_ham.f, fqcomp.f, fqcomp_v2.f, get_col_summary.f, ham_formsol.f
                            ham_varlam.f, recom_chk.f, recomchk_v2.f, sobjbar.f, arnaud_cross_v3_disp.f90

       lte_hydro — det_ed.f, ltepop_wld_v1.f

       misc —  comp_ba_mat_sol.f, extrap_scr.f, guess_dc.f, merge_scr.f, recom_chk_v2.f, rev-rvsg.f
                     solve_ba_mat.f, tlusty_vel.f

       new_main/mod_subs —  adjust_r_grid.f, adjust_r_grid_v2.f, adjust_r_grid_v3.f, adjust_r_grid_v4.f,
                                                comp_j_bank.f, control_variable_mod.f, set_frequency_grid_v2.f,
                                                sub_sob_line.f, sub_sob_line_v2.f, sub_sob_line_v3.f, sub_sob_line_v4.f

        new_main/gam_transport —gamma_emiss_eta_v9.f90, gamray_sub_v3.f90

         /new_main/subs — det_lte_ed.f, do_tau_regrid.f, do_tau_regrid_v2.f, do_vel_regrid.f,
                                        do_vel_regrid_v2.f

         new_main/subs/non_therm  — arnaud_cross.f90, electron_non_therm_spec.f90, get_lelec.f90,
                                                           intsigc.f90, set_xkt_array.f90

         newsubs — omega_gen_v3.f, rdphot_gen_v1.f, rdphot_gen_v2.f

         obs-  cmf_flux_sub_v5.f, comp_ew_abs_ew.f, form_ew.f, obs_frame_vx.f  (x = , 2 ,3 ,4 ,5, 6 7, 8)

         pgplt  — cursor_gaus_fit.f, do_gaus_fit.f, gauss_fit.f, mod_smear_pg.f, set_upper_axis.f, wr_gaus_fit.f

         plane — fgrey_norel_v1.f, jgrey_hub_ddt.f, jgrey_hub_ddt_v2.f, jgrey_hub_ddt_v3.f,
                       jgrey_hub_ddt_v4.f, grey_hub_ddt_v5.f,

         spec_plt — new_plt_rj.f, plt_jh_cur.f, plt_rj.f

         spec_plt/subs —  convolve.f, num_rec.f, raw_subphot_v2.f, rd_xy_data.f, rd_xy_data_usr.f,
                                   uvabs.f, uvabs_v2.f

       subs — comp_j_conv.f, comp_j_conv_v2.f, fg_comp.f, fg_comp_tst.f, formsol.f, gff.f, ins_line_obs_v5.f,
                    jgrey.f, jgrey_with_fvt.f, ltepop_wld_v1.f, ltepop_wld_v2.f, omega_gen_v2.f
                    torscl_v2.f, torscl_v3.f, var_formsol.f, weight.f

      stark — stark/conv_strk_v1.f, stark/conv_strk_v2.f

Some of the following routines were also altered. These have consistencies with the cmf_testing version now called cmf_testing_real8. Routines listed below still need updating and consistency checks.
These include constant
definiotios, and changes made for shock handling.

misc — mod_cool.f  [NEED TO FIX]
             misc   plt_cmf_lum.f   [NEED_TO_FIX]
                        plt_cool_sort.f [NEED TO_FIX}
                        plt_hydro.f
                        tst_ba_mat_sol.f
                        adjust_r_grid_v4.f (needs updating)

Updated WRITE in disp/subs/wr_col_sl.f

new_main/subs/evalse_qwvj_v8.f

newsubs/sub_phot_gen.f

subs/ comp_var_jrec_v2.f

./newsubs/quad_multi_v10.f

subcol_multi_v4.f
   new_main/mod_subs/rd_control_variables.f
   new_main/mod_subs/set_abund_clump.f,

new_main/subs/do_gam_abs_approx_v2.f
new_main/subs/get_edep_shock_power.f
new_main/subs/rd_sn_data.f

new_main/subs/steq_multi_v10.f (remove writes)

new_main/subs/sub_guess_dc.f

./obs/revise_obs_p.f

Fix updated constants: hdkt etc
      new_main/cmfgen.f
      new_main/subs/chg/eval_chg_rates_v3.f
      /new_main/subs/two/two_phot_opac_v3.f
      new_main/subs/two/two_phot_var_opac.f
      newsubs/colcool_sl_v5.f
      subs/colcool_sl_v3.f
      subs/phys_con.f
      subs/prrr_lowt_v2.f, prrr_sl_v6.f, prrr_sl_v5.f
      x_rrr_cool_v6.f
      tools/constants_mod.f

pgplt/gramon_pgplot.f

The following rotines had dLOG_NU replaced by LOG_NU — need to revert back to original notation.
       plane — cmf_formal_rel.f  v2 ad V3 (dLOG_NU — LOG_NU) — fix
                     fg_j_cmf_v11.f plane/mom_j_cmf_v10.f, mom_j_cmf_v11.f
                     v8, v9, v20, mom_j_ddt_v1.f  (v2 …)
                     mom_j_rel_v2.f

./spec_plt/plt_many_sn_spec.f

./spec_plt/plt_scr.f, plt_spec.f

subs/jtrpwgt.f  TRUE — FALSE (change) and ntrpwgt.f

subs/occupation_prob.f {ALTEREED}

subs/regrid_b_on_ne.f
subs/regrid_log_dc_v1.f

unix/dir_acc_pars_gen.f

+++ blas/ddot.f
@@ -1,11 +1,11 @@
-      double precision function ddot(n,dx,incx,dy,incy)
+      REAL(10) function ddot(n,dx,incx,dy,incy)
 c
 c     forms the dot product of two vectors.
 c     uses unrolled loops for increments equal to one.
 c     jack dongarra, linpack, 3/11/78.
 c     modified 12/3/93, array(1) declarations changed to array(*)
 c
-      double precision dx(*),dy(*),dtemp
+      REAL(10) dx(*),dy(*),dtemp
       integer i,incx,incy,ix,iy,m,mp1,n
 c
       ddot = 0.0d0
