+++ blas/drot.f
@@ -4,7 +4,7 @@ c     applies a plane rotation.
 c     jack dongarra, linpack, 3/11/78.
 c     modified 12/3/93, array(1) declarations changed to array(*)
 c
-      double precision dx(*),dy(*),dtemp,c,s
+      REAL(10) dx(*),dy(*),dtemp,c,s
       integer i,incx,incy,ix,iy,n
 c
       if(n.le.0)return
