+++ blas/dswap.f
@@ -5,7 +5,7 @@ c     uses unrolled loops for increments equal one.
 c     jack dongarra, linpack, 3/11/78.
 c     modified 12/3/93, array(1) declarations changed to array(*)
 c
-      double precision dx(*),dy(*),dtemp
+      REAL(10) dx(*),dy(*),dtemp
       integer i,incx,incy,ix,iy,m,mp1,n
 c
       if(n.le.0)return
